requirements.txt
pandas
openpyxl
pymysql
jupyterlab
cryptography
openai
faiss-cpu
tiktoken
python-dotenv
numpy
streamlit
uuid
datetime
